backend/requirements.txt
# Scout backend — Phase 1 + 2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiohttp>=3.9.0
requests>=2.31.0
feedparser>=6.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
playwright>=1.40.0
